+++ FrontendRSReact/src/assets/productpages.css
@@ -92,3 +92,6 @@
   justify-content: space-between;
 }
 
+.product-container {
+  margin-top: 20px; /* Sesuaikan margin atas sesuai kebutuhan */
+}
