+++ FrontendRSReact/src/assets/vendorpages.css
@@ -15,3 +15,5 @@
   align-items: center;
   justify-content: center;
 }
+
+
